keep ringAlarm in entry-point library to preserve callback handle across upgrades
